updated bug in runSuppleClock that was skipping the scanner trigger
The Behavioral version (the behavior version) had been treating a character 'y' or 'n' in a Boolean statement that always evaluated to 'false' the way the code was written.  Took these statements out.  Additionally, the behavior ITI's were not all divided by 10 as intended since we started using this code.